Converting Bowser to Angular Constant & Code Clean-Up
This Bowser Service is great! My team and I use it everywhere! However, we make configurations in our modules based on the browser type. I would be beneficial to us and other users to convert this into a constant. Not only will it be available in BOTH providers and services, but it is better suited as a constant since it does not change while you have an instance of an application running.

In addition, I also did a little clean-up with semi-colons and ===